Compatibility with Numpy 2 (#946)
* empty commit

* compatibility with numpy 2

* changelog

* minor change

* remove openml dep

* doctest

+++ skrub/_to_float32.py
@@ -44,8 +44,8 @@ class ToFloat32(SingleColumnTransformer):
     1    NaN
     2    3.3
     Name: x, dtype: float32
-    >>> float_s[0]
-    1.1
+    >>> float_s[0]            # doctest: +SKIP
+    np.float32(1.1)
 
     Note that a column such as the example above may easily occur as the output
     of ``CleanNullStrings``.
